requirements.txt
google-genai
genai
